deploy/crds: remove resourceHandlingMode=CreateOrUpdate
Setting resourceHandlingMode=CreateOrUpdate was triggering a "kubectl
update" command in resourceHandler.sh. However, the "kubectl" command
has no "update" subcommand. This means that this mode is broken.
Moreover, there are no GitHub issues in eunomia mentioning
"CreateOrUpdate". This means nobody uses that mode.

I guess the intention here was probably to use "kubectl replace", not
"kubectl update". However, seeing as it's not actually used, I propose
to remove it and leave only the "kubectl apply" approach. In theory,
it should be enough. If practice tells us that we really need "kubectl
replace", we can introduce it once again, this time with some concrete
test cases and some solid rationale behind the feature.