chore: create `make generate` target
In Makefile, add new target `generate`, which will run code-generation
subcommands of operator-sdk 0.8.1 (same version as is currently defined
in .travis.yml). The 0.8.1 version of operator-sdk is fairly old, so
it's not easily available in default OS package managers for
contributors to use; therefore, docker is used to easily ensure
the command generates the same code for everybody.